+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # bot-sdk 变更记录
 
+* 增加Display.PushStack指令
+* 增加expectResponse
+* 增加Record.RecordSpeech指令
+* 增加Permission.AskForPermissionsConsent指令
+*** 
 * 增加列表标签Tag
 * 增加BodyTemplate6
 * 增加ListTemplate3